Refactor callbacks, upgrade CDC class

+++ src/classes/msc/mscusb.cpp
@@ -13,7 +13,7 @@ MSCusb::MSCusb()
 bool MSCusb::begin(char* str)
 {
     // Interface number, string index, EP Out & EP In address, EP size
-    uint8_t msc[] = {TUD_MSC_DESCRIPTOR(ifIdx++, 5, _EPNUM_MSC, 0x80 | _EPNUM_MSC, 64)}; // highspeed 512
+    uint8_t msc[] = {TUD_MSC_DESCRIPTOR(ifIdx++, 5, _EPNUM_MSC, (uint8_t)(0x80 | _EPNUM_MSC), 64)}; // highspeed 512
     memcpy(&desc_configuration[total], msc, sizeof(msc));
     total += sizeof(msc);
     count++;
